% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_atbds}
\alias{color_atbds}
\title{The preset color pallete for demo_hbvhcc}
\format{
A named vector for each antigens/antibodies involved
}
\usage{
color_atbds
}
\description{
The preset color pallete for demo_hbvhcc
}
\keyword{datasets}
