% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_hbvhcc}
\alias{demo_hbvhcc}
\title{Original antibody counts data from the article entitled " "
Please cite ""}
\format{
A data frame with 45,502 rows and 36 columns
}
\usage{
demo_hbvhcc
}
\description{
Original antibody counts data from the article entitled " "
Please cite ""
}
\keyword{datasets}
