% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_enrichment}
\alias{estimate_enrichment}
\title{Calculate enrichment for clonotypes and antigens}
\usage{
estimate_enrichment(
  atbd_lst,
  experiment_cells,
  cloneid_vec,
  adjust.method = "BH",
  n_cores = NULL
)
}
\arguments{
\item{atbd_lst, }{calculated by \code{estimate_truecounts}}

\item{experiment_cells}{Vector. Which cells are in the experiment group. Cells not specified here are considered as in the ctrol group.}

\item{cloneid_vec}{Named vector of the clone id for each cell}

\item{adjust.method}{Specify the method for \code{p.adjust}}

\item{n_cores}{By default all the cores available}
}
\value{
A data frame
}
\description{
Calculate enrichment for clonotypes and antigens
}
