% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_truecounts}
\alias{estimate_truecounts}
\title{True-binding cutoffs estimator
The function to estimate cutoffs for identifying true binding counts
from negative controls per antibody.}
\usage{
estimate_truecounts(atbd_mt, experiment_cells, ctrl_p_cutoff = 5e-04)
}
\arguments{
\item{atbd_mt}{Matrix of antibody binding counts from all cells}

\item{experiment_cells}{Vector. Which cells are in the experiment group. Cells not specified here are considered as in the ctrol group.}

\item{ctrl_p_cutoff}{The controled false postive rate}
}
\value{
a named list
}
\description{
True-binding cutoffs estimator
The function to estimate cutoffs for identifying true binding counts
from negative controls per antibody.
}
